requirements.txt
# Core Framework
streamlit>=1.28.0

# AI and ML Libraries
openai>=1.3.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.2

# Vector Database
faiss-cpu>=1.7.4
numpy>=1.24.0

# Document Processing
PyMuPDF>=1.23.0
langchain>=0.0.350
langchain-text-splitters>=0.0.1

# Web Scraping and HTTP
aiohttp>=3.8.0
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0

# Optional: Enhanced Search Providers
duckduckgo-search>=3.9.0
google-api-python-client>=2.100.0

# Development Tools (Optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Security
cryptography>=41.0.0